Auction slot tutorial: improve structure
