+++ docs/tutorials/javascript/trade-on-ledger/index.md
@@ -1,0 +1,9 @@
+---
+metadata:
+  indexPage: true
+---
+# Trade On-Ledger
+
+Use the XRP Ledger's built-in Decentralized Exchange (DEX) to trade assets.
+
+{% child-pages /%}
